+++ src/resolvers/plan.ts
@@ -288,9 +288,9 @@ export const resolvers: Resolvers = {
       }
       return [];
     },
-    sections: async (parent: Plan, _, context: MyContext): Promise<PlanSectionProgress[]> => {
+    versionedSections: async (parent: Plan, _, context: MyContext): Promise<PlanSectionProgress[]> => {
       if (parent?.id) {
-        return await PlanSectionProgress.findByPlanId('plan sections resolver', context, parent.id);
+        return await PlanSectionProgress.findByPlanId('plan versionedSections resolver', context, parent.id);
       }
       return [];
     },
@@ -306,9 +306,9 @@ export const resolvers: Resolvers = {
   },
 
   PlanSearchResult: {
-    sections: async (parent: PlanSearchResult, _, context: MyContext): Promise<PlanSectionProgress[]> => {
+    versionedSections: async (parent: PlanSearchResult, _, context: MyContext): Promise<PlanSectionProgress[]> => {
       if (parent?.id) {
-        return await PlanSectionProgress.findByPlanId('plan sections resolver', context, parent.id);
+        return await PlanSectionProgress.findByPlanId('planSearchresult versionedSections resolver', context, parent.id);
       }
       return [];
     }