+++ src/resolvers/versionedQuestion.ts
@@ -1,13 +1,10 @@
 import { Resolvers } from "../types";
 import { MyContext } from "../context";
 import { VersionedQuestion } from "../models/VersionedQuestion";
-import { VersionedSection } from "../models/VersionedSection";
-import { Section } from "../models/Section";
-import { hasPermissionOnQuestion } from "../services/questionService";
 import { AuthenticationError, ForbiddenError, InternalServerError } from "../utils/graphQLErrors";
 import { VersionedQuestionCondition } from "../models/VersionedQuestionCondition";
 import { prepareObjectForLogs } from "../logger";
-import { isAdmin } from "../services/authService";
+import { isAuthorized } from "../services/authService";
 import { GraphQLError } from "graphql";
 import { normaliseDateTime } from "../utils/helpers";
 
@@ -17,13 +14,25 @@ export const resolvers: Resolvers = {
     publishedQuestions: async (_, { versionedSectionId }, context: MyContext): Promise<VersionedQuestion[]> => {
       const reference = 'publishedQuestions resolver';
       try {
-        if (isAdmin(context?.token)) {
+        if (isAuthorized(context.token)) {
+          return await VersionedQuestion.findByVersionedSectionId(reference, context, versionedSectionId);
+        }
+        // Unauthorized!
+        throw context?.token ? ForbiddenError() : AuthenticationError();
+      } catch (err) {
+        if (err instanceof GraphQLError) throw err;
+
+        context.logger.error(prepareObjectForLogs(err), `Failure in ${reference}`);
+        throw InternalServerError();
+      }
+    },
+
+    publishedQuestion: async (_, { versionedQuestionId }, context: MyContext): Promise<VersionedQuestion> => {
+      const reference = 'publishedQuestion resolver';
+      try {
+        if (isAuthorized(context?.token)) {
           // Grab the versionedSection so we can get the section, and then the templateId
-          const versionedSection = await VersionedSection.findById(reference, context, versionedSectionId);
-          const section = await Section.findById(reference, context, versionedSection.sectionId);
-          if (await hasPermissionOnQuestion(context, section.templateId)) {
-            return await VersionedQuestion.findByVersionedSectionId(reference, context, versionedSectionId);
-          }
+          return await VersionedQuestion.findById(reference, context, versionedQuestionId);
         }
         throw context?.token ? ForbiddenError() : AuthenticationError();
       } catch (err) {